+++ backend/internal/server/apis/schedule/internal/algorithms/round_robin/round_robin.go
@@ -181,17 +181,3 @@ func newGames(season *structures.Season, numberOfGamesPerTeam int) ([]models.Gam
 	return games, nil
 }
 
-func IsEarlyGame(hour, minute int) bool {
-	if hour < 20 {
-		return true
-	}
-	switch hour {
-	case 20:
-		return true
-	case 21:
-		return minute <= 15
-	case 22, 23:
-		return false
-	}
-	return false
-}
