+++ backend/internal/server/server.go
@@ -17,7 +17,7 @@ import (
 	_ "github.com/jak103/powerplay/internal/server/apis/stats"
 	_ "github.com/jak103/powerplay/internal/server/apis/user"
 	_ "github.com/jak103/powerplay/internal/server/apis/groups"
-
+    _ "github.com/jak103/powerplay/internal/server/apis/schedule/games/auto"
 )
 
 func Run() {
